Make the temp stack a bit easier to use
Pushing on to the stack conceptually pushes on the beginning of the
stack (because this is a LIFO stack).  So the 0th index should return
the most recently pushed item.  This commit fixes up the math such that
`temp_stack[0]` and `temp_stack.peek(0)` will return the information for
the most recently pushed item.