Set RAX to RIP
RSI is not always non-zero (I'm not sure why I wrote that in the
comments) and in fact can definitely be zero.  I saw it contain zero on
Linux which meant that `false` was sometimes treated as a heap object.

This commit uses RIP rather than RSI because the instruction pointer
should never be 0.