Flush SP before calling functions that can allocate
Any functions that allocate memory can possibly trigger execution of the
GC.  We need to ensure that the SP is up to date when GC is triggered
because if an object was supposed to be kept alive on the stack and the
SP isn't up to date then the GC will miss marking it.

This commit adds a bunch of SP flushing so that the GC won't miss
objects.